refactor, reformat, and remodel frontend
