+++ frontend/spotted-apple/app/lib/styles/globals.css
