+++ frontend/spotted-apple/features/create-account-form-feature/form.tsx
@@ -3,15 +3,13 @@
 import { ChangeEvent, useState, } from 'react'
 import { FormField, FormFieldError, SubmitButton } from '@/app/lib/forms/components'
 import { ButtonDisplay, InputFields, InputFieldErrorState} from '@/app/lib/forms/types'
-import { SetPasswordFormProps } from '@/app/features/create-set-password-form-feature/types';
+import { CreateAccountFormProps } from '@/features/create-account-form-feature/types';
 
-
-export default function SetPasswordForm({state, formAction}: SetPasswordFormProps) {
-    // TODO: Display one error message at the bottom, not per field
-    const initialErrorState: InputFieldErrorState = {email: true, password: true, confirmPassword: true};
+export default function CreateAccountForm({state, formAction}: CreateAccountFormProps) {
+    const initialErrorState: InputFieldErrorState = {email: true, firstName: true, lastName: true};
     const [intputState, setInput] = useState(state.formData);
     const [errorState, displayError] = useState(initialErrorState) 
-
+    
     const onSubmit = () => { displayError(initialErrorState) }
 
     const onChange = (e: ChangeEvent<HTMLInputElement>) => {
@@ -37,22 +35,27 @@ export default function SetPasswordForm({state, formAction}: SetPasswordFormProp
     const fields: InputFields = [
         {
             id: 'email',
+            placeHolder: 'example@email.com',
             initialValue: intputState && intputState.email ? intputState.email : "",
             onChange: onChange,
         },
         {
-            id: 'password',
+            id: 'firstName',
+            placeHolder: 'First Name',
+            initialValue: intputState && intputState.firstName ? intputState.firstName : "",
             onChange: onChange
         },
         {
-            id: 'confirmPassword',
+            id: 'lastName',
+            placeHolder: 'Last Name',
+            initialValue: intputState && intputState.lastName ? intputState.lastName : "",
             onChange: onChange
         },
     ]
 
-    const formId = 'setPassword';
-    const formHeader = 'Set a password for your account';
-    const buttonValues: ButtonDisplay = {staticDisplay: 'Set Password', pendingDisplay: 'setting password...'}
+    const formId = 'createAccount';
+    const formHeader = 'Create an account with Spotted Apple';
+    const buttonValues: ButtonDisplay = {staticDisplay: 'Create Account', pendingDisplay: 'creating account...'}
 
     return (
         <form id={formId} className={formId} action={formAction} onSubmit={onSubmit}>
@@ -68,7 +71,7 @@ export default function SetPasswordForm({state, formAction}: SetPasswordFormProp
                 </div>
                 ))}
             </ul>
-            <SubmitButton key='setPasswordSubmitButton' {...buttonValues}/>
+            <SubmitButton key='createAccountSubmitButton' {...buttonValues}/>
         </form>
     )
 };
