+++ frontend/spotted-apple/features/home-links-feature/schema.tsx
