+++ frontend/spotted-apple/features/sign-up-feature/actions.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useActionState } from 'react'
-import { SignUpState, SignUpProps, FormAction} from '@/app/features/sign-up-feature/types'
-import { createAccount } from '@/app/features/create-account-form-feature/actions';
-import { setPassword } from '@/app/features/create-set-password-form-feature/actions';
+import { SignUpState, SignUpProps, FormAction} from '@/features/sign-up-feature/types'
+import { createAccount } from '@/features/create-account-form-feature/actions';
+import { setPassword } from '@/features/create-set-password-form-feature/actions';
 
 export default function signUp(state: SignUpState): SignUpProps {
     let formState: SignUpState;
